Reorganize code into packages
