+++ src/main/java/command/AddEventCommand.java
@@ -1,6 +1,15 @@
+package command;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+
+import exception.InvalidDateTimeException;
+import exception.SaveException;
+import component.TaskList;
+import component.Ui;
+import component.Storage;
+import task.Event;
 
 public class AddEventCommand extends Command {
     private final String content;
