+++ src/main/java/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import exception.OutOfListRangeException;
+import exception.SaveException;
+import component.TaskList;
+import component.Ui;
+import component.Storage;
+
 public class DeleteCommand extends Command {
     private final int index;
