+++ src/main/java/command/ExitCommand.java
@@ -1,3 +1,5 @@
+package command;
+
 public class ExitCommand extends Command {
     public ExitCommand() {
