+++ src/main/java/component/Parser.java
@@ -1,7 +1,26 @@
+package component;
+
 import java.util.NoSuchElementException;
 import java.util.Scanner;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import command.AddDeadlineCommand;
+import command.AddEventCommand;
+import command.AddToDoCommand;
+import command.Command;
+import command.DeleteCommand;
+import command.ExitCommand;
+import command.ListCommand;
+import command.MarkCommand;
+import command.UnmarkCommand;
+
+import exception.InvalidIntException;
+import exception.SyntaxException;
+import exception.EmptyFieldException;
+import exception.UnknownCommandException;
+import exception.InvalidDateTimeException;
+import exception.OutOfListRangeException;
 
 public class Parser {
     private static final String CMD_EXIT = "bye";
