+++ src/main/java/exception/InvalidIntException.java
@@ -1,7 +1,9 @@
-public class InvalidDateTimeException extends SpheneException {
+package exception;
+
+public class InvalidIntException extends SpheneException {
     private final String field;
     private final String value;
-    public InvalidDateTimeException(String command, String field, String value) {
+    public InvalidIntException(String command, String field, String value) {
         super(command, "");
         this.field = field;
         this.value = value;
@@ -10,13 +12,12 @@ public class InvalidDateTimeException extends SpheneException {
     @Override
     public String toString() {
         return "Field '" + this.field + "' in command '" + this.getCommand() + "' has value " + this.value
-                + ", not a valid time";
+                + ", not a valid integer";
     }
 
     @Override
     public String getMessage() {
         return "The field '" + this.field + "' in the request '" + this.getCommand() + "' has value " + this.value
-                + ". It is not a valid time!\n"
-                + "Please use the ISO format YYYY-MM-DDTHH:MM:SS.";
+                + ". It is not a valid number!\n";
     }
 }
