+++ src/main/java/exception/SpheneException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public abstract class SpheneException extends Exception {
     private final String command;
     private final String params;
