+++ src/main/java/exception/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class UnknownCommandException extends SpheneException {
     public UnknownCommandException(String command) {
         super(command, null);
