+++ src/main/java/task/Deadline.java
@@ -1,6 +1,7 @@
+package task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Deadline extends Task {
     private final LocalDateTime by;
