+++ src/main/java/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 public abstract class Task {
     private final String content;
     private boolean done;
